+++ Sources/StreamChatUI/ChatMessageList/Attachments/ChatMessageImageGallery+UploadingOverlay.swift
@@ -15,6 +15,10 @@ extension ChatMessageGalleryView {
         
         open var minBottomContainerHeight: CGFloat = 24
 
+        /// The number formatter that converts the uploading progress percentage to textual representation.
+        open lazy var uploadingProgressFormatter: UploadingProgressFormatter =
+            appearance.formatters.uploadingProgress
+
         // MARK: - Subviews
         
         public private(set) lazy var actionButton: AttachmentActionButton = components
@@ -113,9 +117,9 @@ extension ChatMessageGalleryView {
             uploadingProgressLabel.text = content.flatMap {
                 switch $0.state {
                 case let .uploading(progress):
-                    return NumberFormatter.uploadingPercentage.string(from: .init(value: progress))
+                    return uploadingProgressFormatter.format(progress)
                 case .pendingUpload:
-                    return NumberFormatter.uploadingPercentage.string(from: 0)
+                    return uploadingProgressFormatter.format(0)
                 case .uploaded, .unknown:
                     return nil
                 case .uploadingFailed: